Логика для распознавания инвентаря
- Разбил код для работы с моделями на распознавание объектов и классификацию
- Переделал класс Neurolink, чтобы он работал со всеми моделями
- Добавил распознавание чисел
