+++ models/inventory_class_names.txt
@@ -1,0 +1,89 @@
+0
+1
+2
+3
+4
+5
+6
+7
+8
+9
+aequalitas
+aer
+alienis
+aqua
+arbor
+astrum
+auram
+bestia
+caelum
+cognitio
+coralos
+corpus
+desidia
+dreadia
+electrum
+exanimis
+exubitor
+fabrico
+fames
+gelum
+gloria
+granum
+gula
+herba
+humanus
+ignis
+infernus
+instrumentum
+invidia
+ira
+iter
+limus
+lucrum
+lux
+luxuria
+machina
+magneto
+matrix
+messis
+metallum
+meto
+mortuus
+motus
+mru
+nebrisum
+ordo
+pannus
+perditio
+perfodio
+permutatio
+potentia
+praecantatio
+primordium
+radiation
+radio
+sanctus
+sano
+saxum
+sensus
+spiritus
+strontio
+superbia
+tabernus
+telum
+tempestas
+tempus
+tenebrae
+terminus
+terra
+tincturem
+tutamen
+vacuos
+venenum
+vesania
+victus
+vinculum
+vitium
+vitreus
+volatus
