+++ src/logic/onnx_inference/__init__.py
@@ -1,0 +1,2 @@
+from object_detection import OnnxObjectDetection, ObjectPrediction
+from classification import OnnxClassification
